refactor(frontend): split PetDetails into smaller components and hooks (PlacementRequestsSection, PlacementResponseSection, usePlacementInfo, deriveImageUrl); tests passing
